percobaan 2

+++ Jobsheet3/Mahasiswa06.java
@@ -2,5 +2,5 @@ public class Mahasiswa06 {
     public String nim;
     public String nama;
     public String kelas;
-    public double ipk;
+    public float ipk;
 }
